requirements.txt
requests
python-dotenv
sounddevice
scipy
